requirements.txt
allure-pytest==2.14.0
allure-python-commons==2.14.0
annotated-types==0.7.0
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
dotenv==0.9.9
exceptiongroup==1.2.2
execnet==2.1.1
Faker==37.1.0
idna==3.10
iniconfig==2.1.0
packaging==24.2
pluggy==1.5.0
pydantic==2.11.2
pydantic_core==2.33.1
pytest==8.3.5
pytest-xdist==3.6.1
python-dotenv==1.1.0
requests==2.32.3
tomli==2.2.1
tzdata==2025.2
urllib3==2.3.0
